vga text works

+++ drivers/vga_text.h
@@ -43,7 +43,7 @@ static void write_char (char c);
 static void scroll_down ();
 
 extern void write_text_vga (char* str);
-extern void set_cursor (int r, int c);
+extern void set_cursor (u64 r, u64 c);
 extern void clear_screen ();
 extern void fill_screen ();
 extern void set_fg_color (text_color_e);
